refactor: eliminate browser timezone dependencies in trigger schedule
- Remove all browser timezone calculations from execution time computation
- Create dedicated getUserTimezoneCurrentTime function for pure user timezone logic
- Replace formatDateInTimezone with formatUserTimezoneDate to avoid double timezone conversion
- Delete unused timezone-utils.ts and all test files
- Fix core issue where cron '0 22 * * *' showed 6 PM instead of 10 PM due to timezone double conversion
